requirements.txt
streamlit
pandas
nltk
seaborn
matplotlib
bertopic
spacy
PyPDF2
python-docx